+++ src/@types/i18next.d.ts
@@ -1,12 +1,13 @@
 import "i18next";
-import overview from "locales/en/overview.json";
-import scan from "locales/en/scan.json";
-import translation from "locales/en/translation.json";
-import table from "locales/en/table.json";
-import quarantine from "locales/en/quarantine.json";
-import history from "locales/en/history.json";
-import scheduler from "locales/en/scheduler.json";
-import stats from "locales/en/stats.json";
+import overview from "@public/locales/en/overview.json";
+import scan from "@public/locales/en/scan.json";
+import translation from "@public/locales/en/translation.json";
+import table from "@public/locales/en/table.json";
+import quarantine from "@public/locales/en/quarantine.json";
+import history from "@public/locales/en/history.json";
+import scheduler from "@public/locales/en/scheduler.json";
+import stats from "@public/locales/en/stats.json";
+import settings from "@public/locales/en/settings.json"
 
 import { LangCode } from "../i18n/config";
 
@@ -22,6 +23,7 @@ declare module "i18next" {
                history: typeof history;
                scheduler: typeof scheduler;
                stats: typeof stats;
+               settings: typeof settings;
           };
      }
 }