+++ src/lib/constants/settings/tabs.ts
@@ -2,6 +2,7 @@ import { ScanSettingsLoader, GeneralSettingsLoader, AdvancedSettingsLoader, Upda
 import { Braces, Cog, RotateCcw, Search } from "lucide-react";
 import { lazy } from "react";
 import { ISettingsTab } from "@/lib/types";
+import { SettingsTab } from "@/lib/types/enums";
 
 const GeneralSettings = lazy(()=>import("@/contents/settings/general"));
 const AdvancedSettings = lazy(()=>import("@/contents/settings/advanced"));
@@ -10,29 +11,25 @@ const UpdateSettings = lazy(()=>import("@/contents/settings/update"));
 
 export const SETTINGS_TABS: ISettingsTab[] = [
      {
-          page: "general",
-          name: "General",
+          page: SettingsTab.General,
           Icon: Cog,
           Loader: GeneralSettingsLoader,
           LazyComponent: GeneralSettings
      },
      {
-          page: "scan",
-          name: "Scan",
+          page: SettingsTab.Scan,
           Icon: Search,
           Loader: ScanSettingsLoader,
           LazyComponent: ScanSettings
      },
      {
-          page: "advanced",
-          name: "Advanced",
+          page: SettingsTab.Advanced,
           Icon: Braces,
           Loader: AdvancedSettingsLoader,
           LazyComponent: AdvancedSettings
      },
      {
-          page: "update",
-          name: "Update",
+          page: SettingsTab.Update,
           Icon: RotateCcw,
           Loader: UpdateSettingsLoader,
           LazyComponent: UpdateSettings