+++ src/lib/types/index.ts
@@ -2,12 +2,11 @@ import * as z from "zod"
 import { LucideProps } from "lucide-react";
 import { SettingsProps } from "./props";
 import { ExclusionsSchema, SchedulerSchema } from "../schemas";
-import { QuickAccessLink, SidebarLink } from "./enums";
+import { QuickAccessLink, SettingsTab, SidebarLink } from "./enums";
 
 export type ScanType = "" | "main" | "full" | "custom" | "file"
 export type ClamAVState = "checking" | "available" | "missing";
 export type GUIUpdaterStatus = "checking" | "updating" | "needs-update" | "updated" | "failed-check" | "completed" | "failed-update"
-export type SettingsTab = "general" | "advanced" | "scan" | "update"
 export type QuarantineAction = "restore" | "delete"
 export type FsOption = "file" | "folder";
 
@@ -27,8 +26,7 @@ export interface ISidebarItem{
      Icon: React.ForwardRefExoticComponent<Omit<LucideProps, "ref"> & React.RefAttributes<SVGSVGElement>>,
 }
 export interface ISettingsTab{
-     page: SettingsTab,
-     name: string,
+     page: SettingsTab
      Icon: React.ForwardRefExoticComponent<Omit<LucideProps, "ref"> & React.RefAttributes<SVGSVGElement>>,
      Loader: React.FC,
      LazyComponent: React.LazyExoticComponent<React.FC<SettingsProps>>